field length changes, added indexes

+++ commons-app/src/main/java/cz/cuni/mff/xrg/odcs/commons/app/dao/DataObject.java
@@ -13,6 +13,6 @@ public interface DataObject extends Serializable {
     /**
      * @return object's id
      */
-    public abstract int getId();
+    public abstract Long getId();
 
 }
